initial property_list implementation.
Since we have no properties yet, this doesn't actually implement much
it's more of the skeleton that we fill the properties into later.

has_property and get_property are in the property_list implementation
as you can't specialize a member function without explicitly specializing
the containing class, so for buffer properties we do that in here
not in the buffer class.

To handle the pack parameter, we create an iterative addproperty method
that we implement for each base property type using the CREATE macro.

the all_true is required to disambiguate queue constructors, so that
a property_list can only be constructed from a list of classes
derived from the base property detail class.

Property list users are expected to derive this class.