queue: initial implementation of queue
Implement queue get_info and properties support.

We define queue to be derived from the property
list class, as while it would work as a friend
for queues, buffers require templates so it
doesn't work so well there.

Notes:
this currently defaults to host selector (when it
should be default selector, until we have a useful
device compiler.