vec: initial swizzle implementation
This isn't at all spec compliant mainly around the __swizzled_vec__
type and the rules it's meant to follow.

But it adds the APIs and some tests, and is probably half way to
what the proper answer looks like.

This adds src/gen.c which is the C source used to generate the swizzle header files.
I don't think these files need to change that much, so I haven't made it run, and just
checked in the result, we can revisit this later if required.